+++ SOOUM/SOOUM/Base/BaseNavigationViewController.swift
@@ -116,32 +116,6 @@ class BaseNavigationViewController: BaseViewController {
             self.setupNaviBar()
         }
     }
-    
-    func navigationPop(
-        to: UIViewController.Type? = nil,
-        animated: Bool = true,
-        bottomBarHidden: Bool = false,
-        completion: (() -> Void)? = nil
-    ) {
-        CATransaction.begin()
-        CATransaction.setCompletionBlock(completion)
-
-        if let to: UIViewController.Type = to,
-            let viewControllers = self.navigationController?.viewControllers,
-            let destination: UIViewController = viewControllers.last(
-                where: { type(of: $0) == to }
-            ) {
-            destination.hidesBottomBarWhenPushed = bottomBarHidden
-            self.navigationController?.popToViewController(destination, animated: animated)
-        } else {
-            self.navigationController?
-                .viewControllers.dropLast().last?
-                .hidesBottomBarWhenPushed = bottomBarHidden
-            self.navigationController?.popViewController(animated: animated)
-        }
-
-        CATransaction.commit()
-    }
 }
